requirements.txt
openai==1.87.0
python-dotenv==1.0.1
